openai/multi_msgs/requirements.txt
# langchain[llms]
langchain-core
openai
python-dotenv
langchain-openai
# langchain-community